added receipt screen and display information for order
