+++ mobile_app/lib/main_navigation/orders/orders_screen.dart
@@ -27,17 +27,18 @@ class _OrdersScreenState extends State<OrdersScreen> {
   List<Order> completedOrders = [];
   List<Order> inProgressOrders = [];
   List<Order> pickedUpOrders = [];
+  late final StreamSubscription<RemoteMessage> _messageSubscription;
 
   @override
   void initState() {
     super.initState();
 
-    FirebaseMessaging.onMessage.listen((RemoteMessage message) {
-      print("getting something");
-      print(message.data);
+  _messageSubscription =  FirebaseMessaging.onMessage.listen((RemoteMessage message) {
       if (message.data['type'] == 'ORDER_UPDATE') {
         print("automatically updating");
-        _loadOrders();
+        if (mounted) {
+          _loadOrders();
+        }
       }
     });
 
@@ -47,6 +48,7 @@ class _OrdersScreenState extends State<OrdersScreen> {
   @override
   void dispose() {
     _pollingTimer?.cancel();
+    _messageSubscription.cancel();
     super.dispose();
   }
 
@@ -148,7 +150,6 @@ class _OrdersScreenState extends State<OrdersScreen> {
                       child: Text("You haven't placed any orders yet."))
                   : ListView(
                       children: [
-                        
                         if (completedOrders.isNotEmpty) ...[
                           Padding(
                             padding: EdgeInsets.all(verticalSpacing),
@@ -156,10 +157,11 @@ class _OrdersScreenState extends State<OrdersScreen> {
                                 style: TextStyle(
                                     fontSize: 18, fontWeight: FontWeight.bold)),
                           ),
-                          ...completedOrders
-                              .map((order) => Padding(
-                                padding: EdgeInsets.all(horizontalSpacing * 0.5),
-                                child: buildPickupOrderTile(context,order, screenHeight, screenWidth),
+                          ...completedOrders.map((order) => Padding(
+                                padding:
+                                    EdgeInsets.all(horizontalSpacing * 0.5),
+                                child: buildPickupOrderTile(
+                                    context, order, screenHeight, screenWidth),
                               )),
                         ],
                         if (inProgressOrders.isNotEmpty) ...[
@@ -169,10 +171,11 @@ class _OrdersScreenState extends State<OrdersScreen> {
                                 style: TextStyle(
                                     fontSize: 18, fontWeight: FontWeight.bold)),
                           ),
-                          ...inProgressOrders
-                              .map((order) => Padding(
-                                padding: EdgeInsets.all(horizontalSpacing * 0.5),
-                                child: buildProgressOrderTile(order,screenHeight, screenWidth),
+                          ...inProgressOrders.map((order) => Padding(
+                                padding:
+                                    EdgeInsets.all(horizontalSpacing * 0.5),
+                                child: buildProgressOrderTile(
+                                    order, screenHeight, screenWidth),
                               )),
                         ],
                         if (pendingOrders.isNotEmpty) ...[
@@ -192,8 +195,6 @@ class _OrdersScreenState extends State<OrdersScreen> {
                                     order, screenHeight, screenWidth),
                               )),
                         ],
-                        
-                        
                       ],
                     ),
     );
